Now all integer constants are u16.

+++ src/game/paddle.rs
@@ -6,8 +6,8 @@ use super::scene::SCREEN_WIDTH;
 
 use super::ball::DEFAULT_BALL_SPEED;
 
-pub const DEFAULT_PADDLE_HEIGHT: u8 = SCREEN_HEIGHT as u8 / 10;
-pub const DEFAULT_PADDLE_WIDTH: u8 = DEFAULT_PADDLE_HEIGHT / 8;
+pub const DEFAULT_PADDLE_HEIGHT: u16 = SCREEN_HEIGHT / 10;
+pub const DEFAULT_PADDLE_WIDTH: u16 = DEFAULT_PADDLE_HEIGHT / 8;
 pub const DEFAULT_PADDLE_SPEED: f32 = DEFAULT_BALL_SPEED;
 
 pub const DEFAULT_LEFT_CENTER_X: f32 = (PADDLE_MARGIN + DEFAULT_PADDLE_WIDTH / 2) as f32;
@@ -19,13 +19,13 @@ enum Sides {
 
 pub struct Paddle {
     pos: Vec2,
-    width: u8,
-    height: u8,
+    width: u16,
+    height: u16,
     side: Sides,
 }
 
 impl Paddle {
-    pub fn new(pos: Vec2, width: u8, height: u8, side: Sides) -> Self {
+    pub fn new(pos: Vec2, width: u16, height: u16, side: Sides) -> Self {
         Self {
             pos,
             width,
@@ -37,7 +37,7 @@ impl Paddle {
     pub fn default_left_paddle() -> Self {
         Self {
             pos: Vec2::new(DEFAULT_LEFT_CENTER_X, (SCREEN_HEIGHT / 2) as f32),
-            width: PADDLE_MARGIN as u8,
+            width: PADDLE_MARGIN,
             height: DEFAULT_PADDLE_HEIGHT,
             side: Sides::Left,
         }
@@ -49,7 +49,7 @@ impl Paddle {
                 (PADDLE_MARGIN + DEFAULT_PADDLE_WIDTH / 2) as f32,
                 (SCREEN_HEIGHT / 2) as f32,
             ),
-            width: PADDLE_MARGIN as u8,
+            width: PADDLE_MARGIN,
             height: DEFAULT_PADDLE_HEIGHT,
             side: Sides::Right,
         }
